dev-MetOfficeAgent-withinStack: renamed "metoffice" repo to "agent" to prep standardised agent layout
