+++ Agents/MetOfficeAgent/agent/datamodel/iris.py
@@ -1,4 +1,4 @@
-from metoffice.datamodel.observation_types import *
+from agent.datamodel.observation_types import *
 
 ###--- Common Base URLs ---###
 # External ontologies
