+++ Agents/MetOfficeAgent/agent/datamodel/utils.py
@@ -7,7 +7,7 @@
 # of IRIs and PREFIXES while creating SPARQL queries
 
 #import agentlogging
-from metoffice.datamodel.iris import *
+from agent.datamodel.iris import *
 
 # Initialise logger
 #logger = agentlogging.get_logger("prod")
