+++ Agents/MetOfficeAgent/agent/flaskapp/__init__.py
@@ -13,10 +13,10 @@ def create_app(test_config=None):
 
     with app.app_context():
         # Import parts of application
-        import metoffice.flaskapp.home.routes as home
-        import metoffice.flaskapp.inputtasks.routes as inputtasks
+        import agent.flaskapp.home.routes as home
+        import agent.flaskapp.inputtasks.routes as inputtasks
         #TODO: Remove deprecated data retrieval; now handled by FeatureInfoAgent
-        import metoffice.flaskapp.outputtasks.routes as outputtasks
+        import agent.flaskapp.outputtasks.routes as outputtasks
 
         # Register Blueprints
         app.register_blueprint(home.home_bp)
